requirements.txt
pandas == 1.2.4
GoogleNews
newspaper3k